dataTether/requirements.txt
async-timeout==4.0.2
bidict==0.22.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SocketIO==5.3.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
python-engineio==4.4.0
python-socketio==5.8.0
redis==4.5.4
Werkzeug==2.2.3
zipp==3.15.0
